==> experiments/ancient_cn/01_xml_to_txt_n_mysql.ipynb <==
# Ancient Chinese Data
> Process the [ZhWikiSource data](https://dumps.wikimedia.org/zhwikisource/20200301/)

Before running all these code, first we build the docker in the dir docker

## Meta file: index.text

from pathlib import Path
text = Path("/Users/salvor//data/zhwikisource-20200301-pages-articles-multistream.xml")

class XMLGen:
    def __init__(self,path):
        self.f = open(path,mode="r")

    def __iter__(self):
        rt = ""
        while True:
            line = self.f.readline()
            if line:
                if "<page>" in line:
                    rt = ""
                rt+=line
                if "</page>" in line:
                    yield rt
            else:
                break

xg = XMLGen(text)

i = 0
for xml in xg:
    print(xml)
    i+=1
    if i>2:break

from xml.etree import ElementTree as ET

def xmltodict(xml):
    root = ET.fromstring(xml)
    return dict((i.tag,i.text)for i in root.iter())

def process_xml(xml):
    dt = xmltodict(xml)
    

print(xmltodict(xml))

from sqlalchemy import create_engine as ce
import pandas as pd

from base64 import b64encode,b64decode

b64encode("你好".encode())

b64decode(b'5L2g5aW9').decode()

uri = "mysql+pymysql://root:bitnami@localhost/test"
con = ce(uri)

from uuid import uuid4
uuid = uuid4()

def save_text(row):
    id_ = row["id"]
    txt = row["text"]
    uuid = uuid4().hex
    if txt:
        with open(f"/Users/salvor/data/cn_text/{uuid}.txt","w") as f:
            f.write(txt)
    return uuid

from datetime import datetime
def read_time(x):
    try:
        return datetime.strptime(x,"%Y-%m-%dT%H:%M:%SZ")
    except:
        return datetime.now()

create_table = """
CREATE TABLE cn_text
(
    uuid VARCHAR(32) PRIMARY KEY,
    wid INT,
    parentid INT,
    title TEXT CHARACTER SET utf8mb4 COLLATE utf8mb4_unicode_ci,
    ns INT,
    model VARCHAR(20),
    format VARCHAR(20),
    timestamp DATETIME
) CHARACTER SET utf8;
"""

con.execute(create_table)

import sys
xg = XMLGen(text)

i = 0
lines = []
for xml in xg:
    dt = xmltodict(xml)
    i+=1
    lines.append(dt)
    if i%200==199:
        df = pd.DataFrame(lines)
        df["uuid"] = df.apply(save_text,axis=1)
        
        df = df.rename(columns = {"id":"wid"})
        df["timestamp"] = df.timestamp.apply(read_time)
        if "parentid" in df:
            df = df[["uuid","wid","parentid","title","ns","model","format","timestamp"]]
        else:
            df = df[["uuid","wid","title","ns","model","format","timestamp"]]
        
#         df["title"] = df.title.apply(lambda x: b64encode(x.encode()))
        df["wid"] = df.wid.apply(int)
        df = df.set_index("uuid")\
            .to_sql("cn_text",
                    con = con,
                    index = True,
                    if_exists="append")
        lines = []
        sys.stdout.write(f"\r[{i}]\t entries saved")
