==> experiments/ancient_cn/02_title.ipynb <==
# Title
> article titles

from sqlalchemy import create_engine as ce
import pandas as pd
uri = "mysql+pymysql://root:bitnami@localhost/test"
con = ce(uri)

from IPython.display import display

# titles = pd.read_sql("select * from cn_text",con = con)

# # titles["title"] = titles["title"].apply(lambda x: b64decode(x).decode())

# titles.to_csv("/Users/salvor/data/titles.csv",index = False)

titles = pd.read_csv("/Users/salvor/data/titles.csv",)

titles.parentid = titles.parentid.apply(lambda x: int(x) if x>0 else -1)
titles

def preview(uuid):
    print(f"preview: {uuid}.txt")
    print(open(f"/Users/salvor/data/cn_text/{uuid}.txt").read()[:300])

titles.model.value_counts()

for model in titles.model.unique():
    print(model)
    print(titles[titles.model == model][["uuid","wid","title"]].head(5))
    print("="*60)

text = titles[titles.title.str.contains("醒世恆言")][["uuid","title","model","format"]]
text

text = titles[titles.title=="醒世恆言"]
text

preview("1cd5cdbb9c4b4fdf8ecb902eb5b52c40")

text = titles[titles.title=="醒世恆言/第39卷"]
text

text = titles[titles.title=="Category:三言二拍"]
text

preview("bca8bebff75b4f2db8b8f248c07305ec")

text = titles[titles.title=="Category:小说集"]
text

text = titles[titles.title=="Category:古代小说"]
text

preview("299badcb11864551ad32a5dfdbe084c2")

## Title

All category title is unique

titles.title.value_counts()

hence we can search by title

## Name space

quoted from [official explaination](https://wikisource.org/wiki/Wikisource:Namespaces)

>List of namespaces
The two pseudo-namespaces, main namespace and 15 auxiliary namespaces in this project (Meta) are the following (also the variables for them are shown):
```
{{ns:1}}	Talk
{{ns:2}}	User
{{ns:3}}	User talk
{{ns:4}}	Wikisource
{{ns:5}}	Wikisource talk
{{ns:6}}	File
{{ns:7}}	File talk
{{ns:8}}	MediaWiki
{{ns:9}}	MediaWiki talk
{{ns:10}}	Template
{{ns:11}}	Template talk
{{ns:12}}	Help
{{ns:13}}	Help talk
{{ns:14}}	Category
{{ns:15}}	Category talk
{{ns:104}}	Page
{{ns:105}}	Page talk
{{ns:106}}	Index
{{ns:107}}	Index talk
{{ns:108}}	Author
{{ns:109}}	Author talk
{{ns:828}}	Module
{{ns:829}}	Module talk
{{ns:1198}}	Translations
{{ns:1199}}	Translations talk
```

titles.groupby("ns").agg(lambda x:len(x))

for ns in titles.ns.unique():
    subdf = titles[titles.ns == ns]
    display(f"{ns} total lines: {len(subdf)}")
    display(subdf.head()[["uuid","title","ns","model"]])

# proofread-page
preview("00002518b6ea43a086c41c71b8ca92bc")

# 100 portal
preview("00005a91cf4c4d4c98d9226b3912e725")

# 102 Author, usefule, but too few
preview("00047a2b633e45c5ae027f0615e1022a")

# 14 category
preview("0000e44cd4c2488f9398dbc5cab3509e")

# 114 tanslation
preview("023e07a1ced74b4ab3fb34d34fc2c759")

text = titles[titles.title=="詩品"]
text

text = titles[titles.title=="授李彭年等中書舍人制"]
text

preview("c4ca489760cb42c6a5cf64ef0f5258e2")

### Ditch the files we don't  need

# 4 Wikisource
preview("021dbf267d2d4712838acd9c0ab8c705")

# 828 模块:Demo/doc
preview("007ae1fdf6cf499e88ff96cb9551b005")

# 10 template
preview("008855b9e1fa4f8782931946eee9937f")

# 8
preview("00bdab97cb014099b619956fe91f31bc")

# 106 index
preview("004588964eb64509b7db2f0179ff40e9")
preview("001090ee855c437dbe72fbff4db76ae8")

# 12 help
preview("0000e44cd4c2488f9398dbc5cab3509e")

# 114 translation
preview("015525d85a1b4deabeaf51d2ef608877")

# I'm sure the following ns isn't what we want
# pardon my variable name, they are just useless in our task

useless_ns = [4,828,10,8,106,12,2600,6,114]

Remove useless files

import os
for ns in useless_ns:
    subdf = titles[titles.ns==ns]
    for uuid in subdf.uuid:
        os.system(f"rm -rf /Users/salvor/cn_text/{uuid}.txt")
    print(f"ns {ns} files removed")

titles = titles[titles.ns.isin(set(titles.ns.unique()) - set(useless_ns))]

titles

titles.to_csv("/Users/salvor/data/cn_titles.csv",index = False)

!du -sh /Users/salvor/data/cn_titles.csv
