==> experiments/ancient_cn/04_extracted.ipynb <==
# indexing the processed text

[wikiextract](http://medialab.di.unipi.it/wiki/Wikipedia_Extractor) 
```shell
python ~/github/wikiextractor/WikiExtractor.py --processes 4 -ns 0,100,102,14,114 --json zhwikisource-20200301-pages-articles-multistream.xml

python ~/github/wikiextractor/WikiExtractor.py --processes 4 -ns 104 --json zhwikisource-20200301-pages-articles-multistream.xml
```

from pathlib import Path
from glob import glob
import json
import pandas as pd
import os
import sys

HOME = Path(os.environ["HOME"])

DATA = Path(HOME/"data"/"zhwiki")

all_files = glob(str(DATA/"text"/"*"/"*"))
print(len(all_files))
all_files[:5]

def file_to_dict(path):
    with open(path,"r") as f:
        while True:
            line = f.readline()
            if line:
                yield json.loads(line)
            else:
                break

def batch_df(path):
    return pd.DataFrame(list(d for d in file_to_dict(path)))

def hash_name(name):
    return hex(abs(hash(name)))

batch_df(all_files[0])

Create saving destination

!mkdir -p {DATA/"cn_text"}

Saving text instead of json

def save_text(row):
    rid = row["id"]
    hashed = hash_name(row["title"])
    filename = f"{rid}_{hashed}.txt"
    with open(DATA/"cn_text"/filename,"w") as f:
        f.write(row["text"])
    return filename

for path in all_files:
    df = batch_df(path)
    df["filename"] = df.apply(save_text, axis=1)
    df[["id","url","title","filename"]].to_csv(DATA/"zhwikisource_title.csv",
                                               header=False,
                                               index=False,
                                               mode="a", )
    os.system(f"rm {path}")
    sys.stdout.write(f"\r{path}")

for i in list([0,1,2,3,4,5,6,7,8,9,"a","b","c","d","e","f"]):
    for j in list([0,1,2,3,4,5,6,7,8,9,"a","b","c","d","e","f"]):
        path = str(f"/Users/salvor/data/zhwiki/cn_text/*{i}{j}.txt")
        os.system(f"rm -rf {path}")
        sys.stdout.write(f"\r{path}")

df = pd.read_csv("/Users/salvor/data/cn_titles.csv")

for uuid in df.uuid:
    os.system(f"rm -rf /Users/salvor/data/cn_text/{uuid}.txt")
