==> experiments/ancient_cn/07_filter_out_morden_cn_part2.ipynb <==
# Filter out Morden Chinese Text (Part II)
> Create the filter CSV file

> I don't know why, but wikisource is full of legal text, laws, notices etc

import pandas as pd
import numpy as np
import os
import json
import sys
from pathlib import Path

HOME = Path(os.environ["HOME"])

DATA = HOME/"data"
PREVIEW = DATA/"cntext_preview.csv"

df = pd.read_csv(PREVIEW)

df["isy"] = True

df

def stats():
    return pd.DataFrame(df.isy.value_counts())

stats()

# df[df.title.str.contains("司法")].title.sample(50)

df.loc[df.title.str.contains("司法"),"isy"] = False

stats()

def tt_kw(kw):
    df.loc[df.title.str.contains(kw),"isy"] = False
    return stats()

def pv_kw(kw):
    df.loc[df.preview.str.contains(kw),"isy"] = False
    return stats()

def pv_kw_t(kw):
    df.loc[df.preview.str.contains(kw),"isy"] = True
    return stats()

df.loc[df.title.str.contains("人民"),"isy"] = False

stats()

df.loc[df.title.str.contains("通知"),"isy"] = False
stats()

df[df.title.str.contains("国务院")].title.sample(50)

df.loc[df.title.str.contains("国务院"),"isy"] = False
stats()

df[df.title.str.contains("起诉")].title.sample(50)

df.loc[df.title.str.contains("起诉"),"isy"] = False
stats()

df.loc[df.title.str.contains("民國"),"isy"] = False
stats()

df.loc[df.title.str.contains("聯合國"),"isy"] = False
stats()

df.loc[df.title.str.contains("決議"),"isy"] = False
stats()

def sample_search(x):
    sub= df[df.title.str.contains(x)]
    print(sub.isy.value_counts())
    print(sub[sub.isy].sample(10).title)
    return sub.title.sample(20)

sample_search("政府")

def sample_search2(x):
    sub= df[df.preview.str.contains(x)]
    print(sub.isy.value_counts())
    print(sub[sub.isy].sample(7).title)
    return sub.preview.sample(20)

sample_search2("公约")

tt_kw("總統華總一義字")

pv_kw("魯迅"),pv_kw("鲁迅"),pv_kw("戴望舒"),

pv_kw("联合国"),pv_kw("新时代"),pv_kw("新時代")

tt_kw("习近平")

pv_kw("中国人")

pv_kw("国务院"),pv_kw("基本准则")

pv_kw("民國的"),pv_kw("民国的")

pv_kw("共产党")

pv_kw("共和国")

tt_kw("號解釋")

pv_kw("上海市"),pv_kw("吸烟"),pv_kw("条例"),pv_kw("公约")

pv_kw("当事人")

tt_kw("会计准则")

pv_kw("當事人")

pv_kw("中華民國")

pv_kw("中央政府")

tt_kw("中国的")

pv_kw("組織條例")

pv_kw("中華民國")

for i in ["国家","科学","革命","英国","美国","法国","香港","共和国"]:
    df.loc[df.title.str.contains(i),"isy"] = False
stats()

for i in ["毛泽东","蒋介石","云南","广西","北京市","通告","病毒","命令","抗日"]:
    df.loc[df.title.str.contains(i),"isy"] = False
stats()

for i in ["国民党","國民黨","共產黨","共产党"]:
    df.loc[df.title.str.contains(i),"isy"] = False
stats()

df.loc[df.title.str.contains("新闻"),"isy"] = False
df.loc[df.title.str.contains("网络"),"isy"] = False
stats()

df.loc[df.title.str.contains("新聞"),"isy"] = False
stats()

df.loc[df.title.str.contains("细则"),"isy"] = False
stats()

df.loc[df.title.str.contains("管理"),"isy"] = False
stats()

for i in range(61):
    year = f"{i+1960}年"
    df.loc[df.title.str.contains(year),"isy"] = False
stats()

df.loc[df.title.str.contains("规定"),"isy"] = False
stats()

df.loc[df.title.str.contains("办法"),"isy"] = False
stats()

df[df.isy].sample(50)

df.to_csv(DATA/"cntext_rule_sep.csv",index=False)
