==> experiments/ancient_cn/09_top_loss_analysis.ipynb <==
# Top loss analysis

> Analyze the top loss, from BERT classification predicition, against manual rule base label

import pandas as pd
import numpy as np
import os
from path import Path

HOME = Path(os.environ["HOME"])
DATA = HOME/"data"

LOSS_FILE = DATA/"cntext_loss2.csv"

df = pd.read_csv(LOSS_FILE)

df.sample(10)

from ipywidgets import interact,interact_manual
from IPython.display import display

current_list = []

def tag_false(r_list = []):
    global current_list
    global df
    for i in r_list:
        current_list.remove(i)
    df.loc[df.id.isin(r_list),"loss2"] = df.loc[df.id.isin(r_list),"loss2"].apply(lambda x:-abs(x))
    df.loc[df.id.isin(current_list),"isy"] = False

def passneg(row):
    if row["loss"]<0:
        return -abs(row["loss2"])
    else:
        return row["loss2"]

df.loss2 = df.apply(passneg,axis=1)

@interact
def preview_top(page=(1,20,1)):
    global current_list 
    stat = pd.DataFrame(df.isy.value_counts())
    stat = stat.join(pd.DataFrame(df.loss.apply(lambda x:True if x<0 else False).value_counts()))
    stat = stat.join(pd.DataFrame(df.loss2.apply(lambda x:True if x<0 else False).value_counts()))
    display(stat)
    
    subdf = df[df.isy].sort_values(by="loss2",ascending=False).head(page*20).tail(20)
    current_list = list(subdf.id)
    return subdf

tag_false([])

### Set true

@interact_manual
def set_true(kw ="kw",col = ["title","preview"]):
    if kw == "kw": return None
    subdf = df[df[col].str.contains(kw)]
    df.loc[df[col].str.contains(kw),"isy"] = True
    df.loc[df[col].str.contains(kw),"loss2"] = subdf.loss.apply(lambda x: - abs(x))
    return subdf

### Set false

@interact_manual
def set_false(kw ="kw",col = ["title","preview"]):
    if kw == "kw": return None
    subdf = df[df[col].str.contains(kw)]
    df.loc[df[col].str.contains(kw),"isy"] = False
    df.loc[df[col].str.contains(kw),"loss2"] = subdf.loss.apply(lambda x: - abs(x))
    return subdf

list(df[df["preview"].str.contains('的策略')].title)

df.to_csv(DATA/"save_cache_cntext.csv",index = False)

df.to_csv(DATA/"cntext_v3.csv",index = False)
